==> notebooks/02-exploration.Rmd <==
---
title: "02 - Exploration"
output: html_notebook
---

### Imports

```{r}
library(dplyr)
library(ggplot2)
```

```{r}
df <- readRDS("../data/clean-data.rds")
head(df)
```

### Exploratory Data Analysis

Set theme.

```{r}
theme_set(theme_bw())
```

Define color palette.

```{r}
custom_palette <- scale_fill_brewer(palette = "Pastel2")
```

#### Univariate

##### Gender

```{r}
df %>%
  ggplot(aes(Gender)) +
  geom_bar(aes(fill = Gender)) +
  ggtitle("Count by Gender") +
  ylab("Count") +
  custom_palette
```

##### Age

```{r}
df %>%
  ggplot(aes(Age)) +
  geom_bar(aes(fill = Age)) +
  ggtitle("Count by Age") +
  ylab("Count") +
  custom_palette
```

##### Education Level

```{r}
df %>%
  ggplot(aes(Education.Level)) +
  geom_bar(aes(fill = Education.Level)) +
  ggtitle("Count by Education Level") +
  xlab("Education Level") +
  ylab("Count") +
  labs(fill = "Education Level") +
  custom_palette
```

##### Institution Type

```{r}
df %>%
  ggplot(aes(Institution.Type)) +
  geom_bar(aes(fill = Institution.Type)) +
  ggtitle("Count by Institution Type") +
  xlab("Institution Type") +
  ylab("Count") +
  labs(fill = "Institution Type") +
  custom_palette
```

##### IT Student

```{r}
df %>%
  ggplot(aes(IT.Student)) +
  geom_bar(aes(fill = IT.Student)) +
  ggtitle("Count by IT Student") +
  xlab("IT Student") +
  ylab("Count") +
  labs(fill = "IT Student") +
  custom_palette
```

##### Location

```{r}
df %>%
  ggplot(aes(Location)) +
  geom_bar(aes(fill = Location)) +
  ggtitle("Count by Location") +
  xlab("In Town") +
  ylab("Count") +
  labs(fill = "In Town") +
  custom_palette
```

##### Load Shedding

```{r}
df %>%
  ggplot(aes(Load.shedding)) +
  geom_bar(aes(fill = Load.shedding)) +
  ggtitle("Count by Load Shedding") +
  xlab("Load Shedding") +
  ylab("Count") +
  labs(fill = "Load Shedding") +
  custom_palette
```

##### Financial Condition

```{r}
df %>%
  ggplot(aes(Financial.Condition)) +
  geom_bar(aes(fill = Financial.Condition)) +
  ggtitle("Count by Financial Condition") +
  xlab("Financial Condition") +
  ylab("Count") +
  labs(fill = "Financial Condition") +
  custom_palette
```

##### Internet Type

```{r}
df %>%
  ggplot(aes(Internet.Type)) +
  geom_bar(aes(fill = Internet.Type)) +
  ggtitle("Count by Internet Type") +
  xlab("Internet Type") +
  ylab("Count") +
  labs(fill = "Internet Type") +
  custom_palette
```

##### Network Type

```{r}
df %>%
  ggplot(aes(Network.Type)) +
  geom_bar(aes(fill = Network.Type)) +
  ggtitle("Count by Network Type") +
  xlab("Network Type") +
  ylab("Count") +
  labs(fill = "Network Type") +
  custom_palette
```

##### Class Duration

```{r}
df %>%
  ggplot(aes(Class.Duration)) +
  geom_bar(aes(fill = Class.Duration)) +
  ggtitle("Count by Class Duration") +
  xlab("Class Duration") +
  ylab("Count") +
  labs(fill = "Class Duration") +
  custom_palette
```

##### LMS

```{r}
df %>%
  ggplot(aes(Self.Lms)) +
  geom_bar(aes(fill = Self.Lms)) +
  ggtitle("Count by LMS Availability") +
  xlab("LMS Availability") +
  ylab("Count") +
  labs(fill = "LMS Availability") +
  custom_palette
```

##### Device

```{r}
df %>%
  ggplot(aes(Device)) +
  geom_bar(aes(fill = Device)) +
  ggtitle("Count by Device") +
  ylab("Count") +
  custom_palette
```

##### Adaptability Level

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Adaptivity.Level)) +
  ggtitle("Count by Adaptability Level") +
  xlab("Adaptability Level") +
  ylab("Count") +
  labs(fill = "Adaptability Level") +
  custom_palette
```

#### Bivariate

##### Gender

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Gender), position = "dodge") +
  ggtitle("Adaptability Level by Gender") +
  xlab("Adaptability Level") +
  ylab("Count") +
  custom_palette
```

##### Age

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Age), position = "dodge") +
  ggtitle("Adaptability Level by Age") +
  xlab("Adaptability Level") +
  ylab("Count") +
  custom_palette
```

##### Education Level

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Education.Level), position = "dodge") +
  ggtitle("Adaptability Level by Education Level") +
  xlab("Adaptability Level") +
  ylab("Count") +
  labs(fill = "Education Level") +
  custom_palette
```

##### Institution Type

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Institution.Type), position = "dodge") +
  ggtitle("Adaptability Level by Institution Type") +
  xlab("Adaptability Level") +
  ylab("Count") +
  labs(fill = "Institution Type") +
  custom_palette
```

##### IT Student

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = IT.Student), position = "dodge") +
  ggtitle("Adaptability Level by IT Student") +
  xlab("Adaptability Level") +
  ylab("Count") +
  labs(fill = "IT Student") +
  custom_palette
```

##### Location

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Location), position = "dodge") +
  ggtitle("Adaptability Level by Location") +
  xlab("Adaptability Level") +
  ylab("Count") +
  labs(fill = "In Town") +
  custom_palette
```

##### Load Shedding

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Load.shedding), position = "dodge") +
  ggtitle("Adaptability Level by Load Shedding") +
  xlab("Adaptability Level") +
  ylab("Count") +
  labs(fill = "Load Shedding") +
  custom_palette
```

##### Financial Condition

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Financial.Condition), position = "dodge") +
  ggtitle("Adaptability Level by Financial Condition") +
  xlab("Adaptability Level") +
  ylab("Count") +
  labs(fill = "Financial Condition") +
  custom_palette
```

##### Internet Type

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Internet.Type), position = "dodge") +
  ggtitle("Adaptability Level by Internet Type") +
  xlab("Adaptability Level") +
  ylab("Count") +
  labs(fill = "Internet Type") +
  custom_palette
```

##### Network Type

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Network.Type), position = "dodge") +
  ggtitle("Adaptability Level by Network Type") +
  xlab("Adaptability Level") +
  ylab("Count") +
  labs(fill = "Network Type") +
  custom_palette
```

##### Class Duration

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Class.Duration), position = "dodge") +
  ggtitle("Adaptability Level by Class Duration") +
  xlab("Adaptability Level") +
  ylab("Count") +
  labs(fill = "Class Duration") +
  custom_palette
```

##### LMS

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Self.Lms), position = "dodge") +
  ggtitle("Adaptability Level by LMS Availability") +
  xlab("Adaptability Level") +
  ylab("Count") +
  labs(fill = "LMS Availability") +
  custom_palette
```

##### Device

```{r}
df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Device), position = "dodge") +
  ggtitle("Adaptability Level by Device") +
  xlab("Adaptability Level") +
  ylab("Count") +
  custom_palette
```

### Correlation Analysis

#### Chi-Squared Tests

```{r warning=FALSE}
df_p_values <- data.frame(
  p.value = df %>%
    select(-Adaptivity.Level) %>%
    apply(
      MARGIN = 2,
      FUN = function(x) chisq.test(x, df$Adaptivity.Level)$p.value
    ) %>%
    sort() %>%
    unlist()
)

df_p_values
```
