==> notebooks/05-presentation.Rmd <==
---
title: "Presentation"
output: html_notebook
---

Create some visualization for poster.

### Imports

```{r}
library(dplyr)
library(ggplot2)
```

```{r}
df <- readRDS("../data/clean-data.rds")
head(df)
```

### Visualization

Create color variables.

```{r}
color_bg <- "#3D405B"
color_fg <- "#F4F1DE"
color_palette <- c("#E07A5F", "#F2CC8F", "#81B29A")
```

Create custom theme.

```{r}
custom_theme <- theme(
  plot.background = element_rect(fill = color_bg, color = NA),
  panel.border = element_rect(colour = color_fg, fill = NA),
  panel.background = element_rect(fill = color_bg),
  panel.grid.major = element_line(color = color_fg),
  panel.grid.minor = element_line(color = color_fg),
  axis.text = element_text(color = color_fg),
  axis.title = element_text(color = color_fg),
  title = element_text(color = color_fg),
  legend.background = element_rect(fill = color_bg),
  legend.text = element_text(color = color_fg),
  legend.position="bottom"
)
```

Create custom palette.

```{r}
custom_palette <- scale_fill_manual(values = color_palette)
```

#### Count by Adaptability Level

```{r}
p_adapt <- df %>%
  ggplot(aes(Adaptivity.Level)) +
  geom_bar(aes(fill = Adaptivity.Level)) +
  ggtitle("Count by Adaptability Level") +
  xlab("Adaptability Level") +
  ylab("Count") +
  labs(fill = "Adaptability Level") +
  custom_palette +
  custom_theme

p_adapt
```

#### Adaptability Level by Financial Condition

```{r}
df_finance <- df %>%
  group_by(Financial.Condition, Adaptivity.Level) %>%
  count() %>%
  group_by(Financial.Condition) %>%
  mutate(percent = n / sum(n) * 100)

df_finance
```

```{r}
p_finance <- df_finance %>%
  ggplot(aes(Financial.Condition, percent)) +
  geom_bar(aes(fill = Adaptivity.Level), stat = "identity") +
  ggtitle("Adaptability Level by Financial Condition") +
  xlab("Financial Condition") +
  ylab("Percentage") +
  labs(fill = "Adaptability Level") +
  custom_palette +
  custom_theme

p_finance
```



#### Adaptability Level by Institution Type

```{r}
df_institute <- df %>%
  group_by(Institution.Type, Adaptivity.Level) %>%
  count() %>%
  group_by(Institution.Type) %>%
  mutate(percent = round(n / sum(n) * 100))

df_institute
```

```{r}
p_institute <- df_institute %>%
  ggplot(aes(Institution.Type, percent)) +
  geom_bar(aes(fill = Adaptivity.Level), stat = "identity") +
  ggtitle("Adaptability Level by Institution Type") +
  xlab("Institution Type") +
  ylab("Percentage") +
  labs(fill = "Adaptability Level") +
  custom_palette +
  custom_theme

p_institute
```

### Exports

```{r}
ggsave("../visualization/p_adapt.png", p_adapt, width = 4, height = 4)
```

```{r}
ggsave("../visualization/p_finance.png", p_finance, width = 4, height = 4)
```

```{r}
ggsave("../visualization/p_institute.png", p_institute, width = 4, height = 4)
```
